Langchain/12_RealWorldProjects/intelligent_document_processor/requirements.txt
# Azure Dependencies
azure-ai-formrecognizer==3.3.0
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0

# Azure OpenAI
openai==1.6.1
tiktoken==0.5.2

# LangChain
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10

# Vector Store
chromadb==0.4.22

# Web Framework
streamlit==1.29.0
streamlit-extras==0.3.6

# Data Processing
pandas==2.1.4
numpy==1.24.3
pillow==10.1.0
python-docx==1.1.0
openpyxl==3.1.2
pdfplumber==0.10.0
pymupdf==1.23.14

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Monitoring
structlog==23.2.0
